% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_selector.R
\name{input_selecter}
\alias{input_selecter}
\title{Selecteur d'input (drop-down list)}
\usage{
input_selecter(id, label, my_choices)
}
\arguments{
\item{id}{character. The id of the div}

\item{label}{character. The label to be displayed above the selecter}

\item{my_choices}{vector. The list of choices}
}
\value{
a ui element corresponding to selectInput using pre-defined parameters
}
\description{
Selecteur d'input (drop-down list)
}
\examples{
input_selecter(id ="subs",label = "Sélectionnez une substance", my_choices = c("Tout", u_subs))
}
